requirements.txt
fastapi
uvicorn
redis
pyyaml
